+++ src/tightdb/group_shared.hpp
@@ -231,14 +231,13 @@ public:
     bool has_changed();
 
     // The calling thread goes to sleep until the database is changed, or
-    // until wait for change is disabled.
+    // until wait_for_change_release() is called. Return true if the database
+    // has changed, false if it might have. At most one thread may wait in 
+    // wait_for_change() on any single SharedGroup.
     bool wait_for_change();
 
-    // If wait for change is enabled, threads calling wait_for_change() will
-    // suspend until the database is changed or wait for change is disabled.
-    // If wait_for_change is disabled, any thread calling wait_for_change()
-    // return immediately - any thread already waiting will wake up and return
-    void wait_for_change_enable(bool enabled);
+    // release any thread waiting in wait_for_change() on *this* SharedGroup.
+    void wait_for_change_release();
 
     // Transactions:
 
@@ -290,6 +289,7 @@ private:
     util::File::Map<SharedInfo> m_file_map; // Never remapped
     util::File::Map<SharedInfo> m_reader_map;
     std::string m_file_path;
+    bool m_waiting_for_change;
     enum TransactStage {
         transact_Ready,
         transact_Reading,