+++ release_notes.md
@@ -8,6 +8,8 @@
 * Now throwing on overlong table and column names, and on oversized strings and
   binary data values.
 * Fall back to the old query nodes for String as well as int/double/float.
+* Added `SharedGroup::wait_for_change_release()` which allows you to release a
+  thread waiting inside wait_for_change() on a specific SharedGroup instance.
 
 -----------